+++ src/pages/mission/missonCard/MissionContent.tsx
